Add an 'else' clause to the 'if' statement.
To accomplish this, we add keyword tokens for 'if' and 'else', as we
don't want to process them as simple instructions any longer.

We also take a(nother)? page from coffee-script, and use their token
rewriter in order to remove the TERMINATOR token that is normally
emitted between an if clause and its else clause.